pyproject.toml
[project]
name = "box_dump"
version = "1.0.0"
description = "Backup installed packages to JSON/SQLite and visualize drift with NiceGUI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Jesse", email = "jesse@example.com" }]
keywords = ["backup", "packages", "nicegui", "drift"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "peewee>=3.17.0",
    "GitPython>=3.1.40",
    "nicegui>=2.0.0",
]

[project.optional-dependencies]
viewer = ["nicegui>=2.0.0"]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
box-dump = "box_dump.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["box_dump"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["N803", "N806"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"**/viewer.py" = ["F841", "N806"]
